64-bit printing

+++ format.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-char* format_int (char* buffer, int32_t value, uint_fast8_t mincol, uint_fast8_t base);
-char* format_uint (char* buffer, uint32_t value, uint_fast8_t mincol, uint_fast8_t base);
+char* format_int (char* buffer, int64_t value, uint_fast8_t mincol, uint_fast8_t base);
+char* format_uint (char* buffer, uint64_t value, uint_fast8_t mincol, uint_fast8_t base);
 
 #endif
